+++ src/main/java/com/example/calculator/service/CalculatorServiceImpl.java
@@ -1,12 +1,12 @@
 package com.example.calculator.service;
 
-import com.example.calculator.dto.OperationType;
+import com.example.calculator.util.OperationType;
 import org.springframework.stereotype.Service;
 
 @Service
 public class CalculatorServiceImpl implements CalculatorService{
     @Override
-    public Integer binaryOperation(Integer firstOperator, OperationType operationType, Integer secondOperator) {
+    public Long binaryOperation(Long firstOperator, OperationType operationType, Long secondOperator) {
         return operationType.getOperator().apply(firstOperator, secondOperator);
     }
 }
